+++ src/tz_country.rs
@@ -1,6 +1,6 @@
-// tz_country.rs - v10
+// tz_country.rs - v11
 
-fn set_tz_country_10_0(x:&str)->Result<String>{Ok(x.to_string())}
-fn set_tz_country_10_0_check(y:&[u8])->bool{!y.is_empty()}
-struct TZ_COUNTRY_10Inner0{val:u64,name:String}
-impl TZ_COUNTRY_10Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}
+fn do_tz_country_11_0(x:&str)->Result<String>{Ok(x.to_string())}
+fn do_tz_country_11_0_check(y:&[u8])->bool{!y.is_empty()}
+struct TZ_COUNTRY_11Inner0{val:u64,name:String}
+impl TZ_COUNTRY_11Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}